fix(wasm): driver discovery on Windows hosts (O_DIRECTORY workaround)
Under GOOS=js on Windows, os.ReadDir and os.DirFS fail because Go's
js/wasm syscall layer rejects O_DIRECTORY (Node.js doesn't expose
fs.constants.O_DIRECTORY on Windows). This caused listInstalled to
always return empty on Windows hosts.

Split loadDir into build-tagged files: the js variant uses os.Open
(no O_DIRECTORY) + File.ReadDir which works on all hosts. Re-enable
windows-latest in the WASM CI workflow.

Closes #396
