+++ .github/workflows/wasm.yml
@@ -55,10 +55,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        # windows-latest is intentionally excluded: the WASM Windows-host path is
-        # experimental and not yet passing (driver discovery under GOOS=js — see
-        # https://github.com/columnar-tech/dbc/issues/396). Re-add when fixed.
-        os: [ 'ubuntu-latest', 'macos-latest' ]
+        os: [ 'ubuntu-latest', 'macos-latest', 'windows-latest' ]
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
         with:
